# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CryptoStashServer",
    "version": "v3"
  },
  "paths": {
...
    "/Hashrates/{id}": {
      "get": {
        "tags": [
          "Hashrates"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Hashrate"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hashrate"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hashrate"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Hashrates"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Hashrate"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Hashrate"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Hashrate"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Hashrate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Hashrates"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/MiningPools": {
      "get": {
        "tags": [
          "MiningPools"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MiningPool"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MiningPool"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MiningPool"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "MiningPools"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/MiningPool"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MiningPool"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MiningPool"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MiningPool"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MiningPool"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MiningPool"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MiningPool"
                }
              }
            }
          }
        }
      }
    },
    "/MiningPools/{id}": {
      "get": {
        "tags": [
          "MiningPools"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MiningPool"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MiningPool"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MiningPool"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "MiningPools"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/MiningPool"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MiningPool"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MiningPool"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MiningPool"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "MiningPools"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/Workers/{id}": {
      "get": {
        "tags": [
          "Workers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Worker"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Worker"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Worker"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Workers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Worker"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Worker"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Worker"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Worker"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Workers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Coin": {
        "type": "object",
        "properties": {
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "ticker": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Hashrate": {
        "type": "object",
        "properties": {
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "current": {
            "type": "integer",
            "format": "int32"
          },
          "average": {
            "type": "integer",
            "format": "int32"
          },
          "reported": {
            "type": "integer",
            "format": "int32"
          },
          "worker": {
            "$ref": "#/components/schemas/Worker"
          }
        },
        "additionalProperties": false
      },
      "MiningPool": {
        "type": "object",
        "properties": {
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "poolBalances": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PoolBalance"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PoolBalance": {
        "type": "object",
        "properties": {
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "current": {
            "type": "number",
            "format": "double"
          },
          "miningPoolId": {
            "type": "integer",
            "format": "int32"
          },
          "miningPool": {
            "$ref": "#/components/schemas/MiningPool"
          },
          "walletId": {
            "type": "integer",
            "format": "int32"
          },
          "wallet": {
            "$ref": "#/components/schemas/Wallet"
          }
        },
        "additionalProperties": false
      },
      "Wallet": {
        "type": "object",
        "properties": {
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "coinId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "coin": {
            "$ref": "#/components/schemas/Coin"
          },
          "workers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Worker"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Worker": {
        "type": "object",
        "properties": {
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "lastModified": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "wallet": {
            "$ref": "#/components/schemas/Wallet"
          },
          "miningPool": {
            "$ref": "#/components/schemas/MiningPool"
          }
        },
        "additionalProperties": false
      }
    }
  }
}